Make mysql tls parameter configurable
Allow the user to specify the tls setting for the
mysql connection. An example would be setting tls
to "preferred", or "true". This is separate to
the tlsConfig config parameter, which is used to
set up a custom tls config, where tls key/certs
are specified.

See the tls parameter section in the below link:
https://pkg.go.dev/github.com/go-sql-driver/mysql#section-readme

Signed-off-by: Edward Broeder <eddie.broeder@intel.com>