chore: checkout code first in CI
setup-go has built-in functionality for caching and restoring go modules and build outputs. The action defaults to
search for the dependency file - go.sum in the repository root, and uses its hash as a part of the cache key.
To use the cache, we should first checkout code. Otherwise setup-go can't find go.sum and use/restore cache.
Related: https://github.com/actions/setup-go#caching-dependency-files-and-build-outputs

Signed-off-by: Gaius <gaius.qi@gmail.com>