feat: replace keys with scan in redis client
It's not recommended to use the KEYS prefix:* command to get all the keys in a Redis instance,
especially in production environments, because it can be a slow and resource-intensive operation
that can impact the performance of the Redis instance.

Instead, you can iterate over Redis keys that match some pattern using the SCAN command,
refer to https://redis.uptrace.dev/guide/get-all-keys.html.

Signed-off-by: Gaius <gaius.qi@gmail.com>